NumericalTools

Documentation for NumericalTools.jl.

This package offers a collection of numerical tools that are useful for daily scientific research. Some of these tools, including geomspace and linspace, were inspired by the popular Python library, numpy.

Index

All the implemented tools are listed below:

NumericalTools.geomspace

NumericalTools.linspace

NumericalTools.sqrtm1

Tools

NumericalTools.geomspace — Function

geomspace(start, stop, num=50; endpoint=true)

Generate a Vector of geometric sequence,

\[v_i = a \left(\frac{b}{a}\right)^{(i - 1)/n}\]

for $i = 1, 2, ⋯, N$, where $N =$ num, $a =$ start, $b =$ stop, and $n =$ num $- 1$ if endpoint = true, otherwise $n =$ num.

Arguments

start::Number: The starting value of the sequence.

stop::Number: The final value of the sequence.

num::Integer=50: Number of samples to generate.

Keywords

endpoint::Bool=true: If true, stop is the last sample. Otherwise, it is not included.

Example

julia> geomspace(1, 1e4, 5)
5-element Vector{Float64}:
     1.0
    10.0
   100.0
  1000.0
 10000.0

NumericalTools.linspace — Function

linspace(start, stop, num=50; endpoint=true)

Generate a Vector of arithmetic sequence,

\[v_i = a + (i - 1) \frac{b - a}{n}\]

for $i = 1, 2, ⋯, N$, where $N =$ num, $a =$ start, $b =$ stop, and $n =$ num $- 1$ if endpoint = true, otherwise $n =$ num.

Arguments

start::Number: The starting value of the sequence.

stop::Number: The final value of the sequence.

num::Integer=50: Number of samples to generate.

Keywords

endpoint::Bool=true: If true, stop is the last sample. Otherwise, it is not included.

Example

julia> linspace(1.0, 5.0, 5)
5-element Vector{Float64}:
 1.0
 2.0
 3.0
 4.0
 5.0

NumericalTools.sqrtm1 — Method

sqrtm1(x)

This function calculates

\[ \text{sqrtm1}(x) = \sqrt{1 + x} - 1\]

for $x > -1$, and takes care of small x.

Example

julia> sqrt(1 + 1e-16) - 1
0.0

julia> sqrtm1(1e-16)
5.0e-17