NumericalTools

Documentation for NumericalTools.jl.

This package offers a collection of numerical tools that are useful for daily scientific research. Some of these tools, including geomspace and linspace, were inspired by the popular Python library, numpy.

Index

All the implemented tools are listed below:

NumericalTools.geomspace

NumericalTools.linspace

NumericalTools.loginterpolator

NumericalTools.sqrtm1

Tools

NumericalTools.geomspace — Function

geomspace(start, stop, num=50; endpoint=true)

Generate a Vector of geometric sequence,

\[v_i = a \left(\frac{b}{a}\right)^{(i - 1)/n}\]

for $i = 1, 2, ⋯, N$, where $N =$ num, $a =$ start, $b =$ stop, and $n =$ num $- 1$ if endpoint = true, otherwise $n =$ num.

Arguments

start::Number: The starting value of the sequence.

stop::Number: The final value of the sequence.

num::Integer=50: Number of samples to generate.

Keywords

endpoint::Bool=true: If true, stop is the last sample. Otherwise, it is not included.

Example

julia> geomspace(1, 1e4, 5)
5-element Vector{Float64}:
     1.0
    10.0
   100.0
  1000.0
 10000.0

NumericalTools.linspace — Function

linspace(start, stop, num=50; endpoint=true)

Generate a Vector of arithmetic sequence,

\[v_i = a + (i - 1) \frac{b - a}{n}\]

for $i = 1, 2, ⋯, N$, where $N =$ num, $a =$ start, $b =$ stop, and $n =$ num $- 1$ if endpoint = true, otherwise $n =$ num.

Arguments

start::Number: The starting value of the sequence.

stop::Number: The final value of the sequence.

num::Integer=50: Number of samples to generate.

Keywords

endpoint::Bool=true: If true, stop is the last sample. Otherwise, it is not included.

Example

julia> linspace(1.0, 5.0, 5)
5-element Vector{Float64}:
 1.0
 2.0
 3.0
 4.0
 5.0

NumericalTools.loginterpolator — Method

loginterpolator(x, y; method="loglog", extrapolation_bc=nothing)

Log-log and semi-log interpolator. Return an interpolator function.

Arguments

x::AbstractVector: x array.

y::AbstractVector: y array.

Keywords

method="loglog": {"loglog", "xlog", "ylog"}, optional. Choose which axis to set in log scale.

extrapolation_bc=nothing: Pass to linear_interpolation.

Example

julia> x = geomspace(1e-1, 10); y = x.^2; itp = loginterpolator(x, y);

julia> itp(1)
1.0

julia> itp(0.5)
0.25

julia> itp(5)
24.999999999999996

NumericalTools.sqrtm1 — Method

sqrtm1(x)
sqrtm1(x, a)

These functions calculate

\[ \text{sqrtm1}(x) = \sqrt{1 + x} - 1,\]

for $x > -1$, where $x$ is dimensionless, and

\[ \text{sqrtm1}(x, a) = \sqrt{a^2 + x} - a,\]

where $x$ and $a^2$ have the same dimension. They are designed to handle situations where the absolute value of x is much smaller than 1 and a, where a is positive.

Example

julia> sqrt(1 + 1e-16) - 1
0.0

julia> sqrtm1(1e-16)
5.0e-17

julia> sqrt(1e16^2 + 1) - 1e16
0.0

julia> sqrtm1(1, 1e16)
5.0e-17

julia> m = 1; v = 1e-10; p = m*v; T = 1/2*m*v^2
5.0000000000000005e-21

julia> sqrt(p^2 + m^2) - m
0.0

julia> sqrtm1(p^2, m)
5.0000000000000005e-21